Extract most of the repo info check into a validator
